enabled logging of unknown packets and when packets were not fully read

+++ chat-server/src/com/aionemu/chatserver/common/netty/BaseServerPacket.java
@@ -10,7 +10,7 @@ public abstract class BaseServerPacket extends AbstractPacket {
 	/**
 	 * @param opCode
 	 */
-	public BaseServerPacket(int opCode) {
+	public BaseServerPacket(byte opCode) {
 		super(opCode);
 	}
