+++ chat-server/src/com/aionemu/chatserver/network/aion/AbstractServerPacket.java
@@ -13,7 +13,7 @@ public abstract class AbstractServerPacket extends BaseServerPacket {
 	/**
 	 * @param opCode
 	 */
-	public AbstractServerPacket(int opCode) {
+	public AbstractServerPacket(byte opCode) {
 		super(opCode);
 	}
